+++ grako/__init__.py
@@ -5,14 +5,14 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from grako._config import __version__
 from grako._config import __toolname__
-from grako.tool import compile, parse, gencode
+from grako.tool import compile, parse, to_python_sourcecode
 from grako.tool import main
 
 assert __version__
 assert __toolname__
 assert compile
 assert parse
-assert gencode
+assert to_python_sourcecode
 
 
 if __name__ == '__main__':